+++ app/src/pages/AddressList.tsx
@@ -2,6 +2,8 @@ import {
   IonButton,
   IonButtons,
   IonContent,
+  IonFab,
+  IonFabButton,
   IonHeader,
   IonIcon,
   IonItem,
@@ -21,7 +23,7 @@ import {
 import { useContext, useEffect, useMemo, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { StoreContext } from "../stores/Store";
-import { openOutline, searchOutline } from "ionicons/icons";
+import { mapOutline, openOutline, searchOutline, shield } from "ionicons/icons";
 import IsPrivate from "../utils/is_private";
 import OverlayPaused from "../components/OverlayPaused";
 import OverlayAppDisabled from "../components/OverlayChainAppDisabled";
@@ -34,6 +36,7 @@ import wrapIndex from "../utils/wrap_index";
 import { useDebounce } from "@uidotdev/usehooks";
 import IsPaused from "../utils/is_paused";
 import { Sleep } from "../utils/sleep";
+import RouteMapPopup from "../components/RouteMap/RouteMapPopup";
 
 export default function AddressList() {
   const {
@@ -52,6 +55,7 @@ export default function AddressList() {
   } = useContext(StoreContext);
   const { t } = useTranslation();
   const headerSheetModal = useRef<HTMLIonModalElement>(null);
+  const refRouteMapPopup = useRef<HTMLIonModalElement>(null);
 
   const headerKey = "addressList";
 
@@ -139,6 +143,10 @@ export default function AddressList() {
     Promise.all([refreshPromise, sleepPromise]).then(() => {
       e.detail.complete();
     });
+  }
+
+  function onClickFab() {
+    refRouteMapPopup.current?.present();
   }
 
   return (
@@ -263,6 +271,14 @@ export default function AddressList() {
               return <AddressListItem {...props} key={props.number} />;
             })}
           </IonList>
+          {chain && authUser ? (
+            <RouteMapPopup
+              chain={chain}
+              modal={refRouteMapPopup}
+              authUserUID={authUser.uid}
+              isChainAdmin={isChainAdmin}
+            />
+          ) : null}
           {isChainAdmin ? (
             <EditHeaders
               modal={headerSheetModal}
@@ -278,6 +294,25 @@ export default function AddressList() {
             className="tw-w-full -tw-mb-2 tw-invert-[60%] tw-overflow-hidden tw-stroke-text dark:tw-stroke-light-tint"
           />
         </div>
+        {isChainAdmin || chain?.allow_map ? (
+          <IonFab slot="fixed" horizontal="end" vertical="bottom">
+            <IonFabButton
+              color={chain?.allow_map && isThemeDefault ? "danger" : ""}
+              onClick={onClickFab}
+              style={
+                chain?.allow_map
+                  ? {}
+                  : {
+                      "--background": "var(--ion-color-blue-contrast)",
+                      "--background-hover": "var(--ion-color-light-shade)",
+                      "--color": "var(--ion-color-blue)",
+                    }
+              }
+            >
+              <IonIcon icon={mapOutline} />
+            </IonFabButton>
+          </IonFab>
+        ) : null}
       </IonContent>
     </IonPage>
   );